lesson-04/requirements.txt
Django==1.4.3
Werkzeug==0.8.3
django-annoying==0.7.6
django-extensions==1.0.2
gunicorn==0.6.1
